fix: resolve all clippy warnings across workspace
- Fix 60 empty-line-after-doc-comment warnings across filter files
- Fix 8 doc-list-item-without-indentation warnings in enhancement/codegen docs
- Fix collapsible-if, div_ceil, redundant-closure via cargo clippy --fix
- Replace manual slice copies with copy_from_slice in box_blur, healing_brush
- Convert indexed loops to iterators in color_lut, invert_v2, subgraph
- Simplify boolean expression in v2_adapters
- Fix struct-update-no-effect in codegen pipeline_write template
- Add #[allow(clippy::type_complexity)] on inherently complex fn-pointer fields
- Add #[allow(clippy::too_many_arguments)] on stroke rasterization function
- Wrap generated CLI dispatch in module with targeted clippy allows

+++ crates/rasmcore-image/src/domain/filters/mask/luminance_range.rs
@@ -35,8 +35,8 @@ pub fn mask_luminance_range(
     };
 
     let n = (info.width as usize) * (info.height as usize);
-    let black_val = (black.clamp(0.0, 1.0) * 255.0) as f32;
-    let white_val = (white.clamp(0.0, 1.0) * 255.0).max(black_val) as f32;
+    let black_val = black.clamp(0.0, 1.0) * 255.0;
+    let white_val = (white.clamp(0.0, 1.0) * 255.0).max(black_val);
     let feath = feather.clamp(0.0, 128.0);
 
     let mut mask = Vec::with_capacity(n);
